requirements.txt
scrapy
telegram-send
coloredlogs
pyyaml